requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
mysql-connector-python==9.0.0
gunicorn==21.2.0
